django_api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
python-decouple==3.8
psycopg2-binary==2.9.7
requests==2.31.0
cryptography==41.0.7
django-extensions==3.2.3
boto3==1.29.7
gunicorn==21.2.0
whitenoise==6.6.0
